==> Slides/03-Introducao-ao-R-Introducao-Escrita-de-Funcoes-slides.Rmd <==
---
title: "Introdução ao R para Ecólogos"
subtitle: "⚔<br/>Introdução a escrita de funções em R"
author: "Vitor Borges"
institute: "PPGBIO-UNIRIO"
date: "2021/05/24 (updated: `r Sys.Date()`)"
output:
  xaringan::moon_reader:
    seal: false
    lib_dir: libs
    nature:
      slideNumberFormat: ""
      highlightStyle: github
      highlightLines: true
      countIncrementalSlides: false
---

layout: true
name: basic
background-image: url("/home/cla/Documentos/Vitor/Laboratórios/UNIRIO/Disciplinas/Analise-de-Dados-Ecologicos/libs/images/AnalisaR_teste1.png")
background-position: 0% 100%
background-size: 20%
class: inverse

```{r setup, include=FALSE}
options(htmltools.dir.version = FALSE)
library(tidyverse)
df <- read.table("/home/cla/Documentos/Vitor/Laboratórios/UNIRIO/Disciplinas/Introdução-ao-R-para-Ecólogos/IRE-21/Dados/df.txt", header = TRUE)
xaringanExtra::use_xaringan_extra()
xaringanExtra::use_clipboard()
```

---
name: Title
background-image: url("/home/cla/Documentos/Vitor/Laboratórios/UNIRIO/Disciplinas/Analise-de-Dados-Ecologicos/libs/images/AnalisaR_teste1.png")
background-position: 0% 100%
background-size: 20%
class: inverse, center, middle, title-slide

# .blue[Introdução ao R para Ecólogos]

## .blue[⚔<br/>Introdução a escrita de funções em R]

### .blue[Vitor Borges-Júnior]

### .blue[PPGBIO - UNIRIO]

### .blue[2021/06/08]

---

# .blue[**Funções**]
 

<BR/>

- Toda .red[ação] no `R` se dá através das .red[funções]<br/>
<BR/>
- Funcionam como .red["caixas pretas"]<br/>
<BR/>

```{r caixa preta, echo = FALSE, fig.align='center'}
knitr::include_graphics("https://designshack.net/wp-content/uploads/blackbox-3.jpg")
```

???
Mostrar exemplo de funções quando falar em caixa preta

---

# .blue[**Funções**]
 

<BR/>

- Muitas delas já estão programadas direto na `base` do R...<br/>
<BR/>
- Outras estão programadas em .red[pacotes]<br/>
<BR/>
- Outras serão programadas por .red[você]<br/>
<BR/>
- .red[Por que] escrever funções?<br/>
<BR/>

???


---

# .blue[**Benefícios de escrever funções**]

<BR/>

- .red[Eliminar repetição] nos scripts<br/>
<BR/>
- .red[Minimizar] a digitação<br/>
<BR/>
- .red[Evitar erros copia/cola]<br/>
<BR/>
- Facilita a .red[reutilização] do código em diferentes projetos<br/>
<BR/>
- Escrever funções para atender suas .red[necessidades]<br/>
<BR/>

???

---

# .blue[**Sintaxe de uma função**]

<BR/>

- Usamos .red[`function() {}`] para criar uma função

```{r primeira função, eval = FALSE, echo = TRUE}
minha_funcao <- function(arg1, arg2){
  # Faça alguma coisa
}
```
--
- .red[Chamada]
```{r, eval = FALSE}
function(arg1, arg2) 
```
--
- .red[Algoritmo]
```{r, eval = FALSE}
{
  # Faça alguma coisa!
}
```

???

Mostrar exemplos de argumentos

---

# .blue[**Exemplo simples**]

<BR/>

- Função .red[`soma()`]

```{r função soma}
# função que soma dois valores
soma <- function(x, y) {
  print(x + y)
}

# usando a função soma
soma(2, 5)
```

???

---

# .blue[**Outro exemplo simples**] 

<BR/>

- Função .red[`razao()`] - calcula a razão entre dois valores:<br/>
<BR/>

```{r função razao() 1, eval = TRUE, echo = TRUE}
razao <- function(valor1, valor2){
  
  print(valor1/valor2)
  
}

# Uso da função
razao(valor1 = 10, valor2 = 5)
```

???


---

# .blue[**Outro exemplo simples**] 

<BR/>

* Podemos .red[modificar a ordem] em que alimentamos os .red[argumentos], ou mesmo .red[suprimir] o nome dos argumentos (mas mantendo a ordem original)<br/>
<BR/>
```{r função razao() 2, eval = FALSE, echo = TRUE}
# forma mais 'correta'
razao(valor1 = 10, valor2 = 5)

# mesmo resultado
razao(valor2 = 5, valor1 = 10)

# não é necessário forcencer os nomes dos argumentos
razao(10, 5)
```

---

# .blue[**Argumentos *default* **] 

* Uma função pode ter valores padrão (.red[*default*]) - na ausência de uma determinação do usuário, ele vai assumir que tudo deve ser feito utilizando o valor padrão:<br/>
<BR/>
```{r argumento default, eval = TRUE, echo = TRUE}
# vamos dar um valor padrão para um dos tratamentos
razao <- function(valor1, valor2 = 100){
  print(valor1/valor2)
  }

# se não especificarmos o argumento valor2
razao(valor1 = 10)

# e se especificarmos o argumento valor2?
razao(valor1 = 10, valor2 = 5)
```


???
Exemplo de argumento default numérico. Mas eles também podem ser `NULL`, lógicos ou categóricos. Veremos exemplos em breve.

---

# .blue[**Objetos dentro da função**]

* Também podemos criar .red[objetos dentro da função]<br/>
<BR/>
* Estes .red[objetos só existirão dentro da função] que estamos criando<br/> 
<BR/>
* Lembrem-se que ao criar um objeto, o padrão do R é .red[não mostrar o resultado]<br/>
<BR/>
```{r criar objeto dentro da função, eval = TRUE, echo = TRUE}
# salvando resultado do cálculo em um objeto dentro da função
razao <- function(valor1, valor2 = 100){
  resposta <- valor1/valor2
}

# usando a função - cade o resultado?
razao(valor1 = 10)

# obrigando o R a mostrar o resultado
(razao(valor1 = 10, valor2 = 5))
```

???
Introduzir regras de scopo.

---

# .blue[**Outra forma de obter o resultado**] 

<BR/>

- Podemos .red[salvar] o resultado da execução da função em um .red[novo objeto] para então acessar o resultado nele armazenado<br/>
<BR/>
```{r obtendo resultado, eval = TRUE, echo = TRUE}
razao <- function(valor1, valor2){
  resposta <- valor1/valor2
}

# outra forma de obter o resultado
cade <- razao(valor1 = 10, valor2 = 5)
cade
```

---

# .blue[**Forma adequada de obter o resultado**]

<BR/>

* Forma mais adequada para contornar este problema é o uso da função .red[`print()`]:<br/>
<BR/>
```{r usanod print(), eval = TRUE, echo = TRUE}
# adicionando a função print
razao <- function(valor1, valor2){
  resposta <- valor1/valor2
  print(resposta)
}

# usando função
razao(valor1 = 10, valor2 = 5)
```

---

# .blue[**Forma adequada de obter o resultado** ]

<BR/>

- Ou o controlador de fluxo .red[`return()`]<br/>
<BR/>
```{r usando return(), eva = TRUE, echo = TRUE}
# adicionando o controlador de fluxo return
razao <- function(valor1, valor2){
  resposta <- valor1/valor2
  return(resposta)
}

# usando função
razao(valor1 = 10, valor2 = 5)
```

---

# .blue[**Função `erro_padrao()`**]

O R não possui uma função capaz de calcular o erro padrão da média de um vetor. Mas com suas novas habilidades você resolverá esse problema criando a função .red[`erro_padrão()`]
- Dica: a fórmula para o erro padrão da média é $\frac{dev.\ pad.}{\sqrt{n}}$   
    1. Calcular o .red[desvio padrão]. Você pode usar, por exemplo a coluna `area` do data frame `df` dentro da função `sd()`; 
    2. Determinar o .red[comprimento] de `area` usando a função `length()` e calcular a sua raiz quadrada usando a função `sqrt()`;  
    3. Dividir o resultado do passo 1 pelo passo 2;  
    4. Insira as linhas de código que você gerou, seu algorítmo para calcular o erro padrão, dentro do corpo da sua nova função:

```{r erro_padrao, echo = TRUE, eval = FALSE}
erro_padrao <- function(valor){
      insira seu código aqui
    }
```
    
.center[Lembre-se de utilizar as funções `print()` ou `return()` para que a função retorne o resultado!]

???
Exercício do notebook 03

---

# .blue[**Convertendo scripts em funções**]

<BR/>

1. Prepare um .red["modelo"]<br/>
<BR/>
2. .red[Cole] o script no interior do "modelo"<br/>
<BR/>
3. Escolha os .red[argumentos]<br/>
<BR/>
4. .red[Substitua] os objetos internos pelos .red[argumentos equivalentes]<br/>
<BR/> 
5. Use .red[nomes generalistas] para as variáveis internas a função<br/>
<BR/> 
6. .red[Teste] a função!

???


---

# .blue[**Funções mais complexas: arremessos de dado**]

* Podemos avançar um pouco mais e tentar trabalhar uma ideia um pouco mais complexa de função: .red[arremessar um dado]!

* O primeiro passo para criar uma função para arremessar um dado será .red[criar] nosso dado: 

```{r criar dado, eval = TRUE, echo = TRUE}
dado <- 1:6
```

* Vamos simular um arremesso usando a função .red[`sample()`]:

```{r simular 1 arremesso, eval = TRUE, echo = TRUE}
# como é uma aleatorização, precisamos usar o set.seed
set.seed(0)

# queremos apenas um arremeso da moeda
numero_de_arremessos <- 1
sample(x = dado, size = numero_de_arremessos)
```

---

# .blue[**Arremessos de dado**]

* Já temos o formato que a função `sample()` deve assumir no corpo da função. Agora, vamos .red[testá-la] com 10 arremessos:

```{r simular 10 arremessos 1, eval = FALSE, echo = TRUE}
numero_de_arremessos <- 10
sample(x = dado, size = numero_de_arremessos)
```
Error in sample.int(length(x), size, replace, prob) : impossível tomar uma amostra maior que a população quando 'replace = FALSE'

--
count: false

- A função `sample()` pode realizar amostragens com ou sem reposição. No nosso exemplo da moeda precisamos informar a `sample()` que desejamos uma .red[amostragem com reposição] adicionando o argumento .red[`replace = TRUE`]

```{r simular 10 arremessos 2, eval = TRUE, echo = TRUE}
numero_de_arremessos <- 10
sample(x = dado, size = numero_de_arremessos, replace = TRUE)
```

???


---

# .blue[**Arremessos de dado**]

* Assim, fica mais fácil desenhar a função, pois .red[sabemos quais os argumentos] da função que estamos criando que deverão ser manipulados!

```{r função arremecar_dado, eval = TRUE, echo = TRUE}
# criando função
arremecar_dado <- function(numero_de_arremessos) {
  dado <- 1:6
  resultado <- sample(x = dado, size = numero_de_arremessos,
         replace = TRUE)
  return(resultado)
}
# lancando dado 33 vezes
set.seed(0)
arremecar_dado(numero_de_arremessos = 33)
```


???
Introduzir a função `table()` para mostrar o resultado em formato de tabela de frequencias

---

# .blue[**Arremessos de dado viciado**]

* Podemos .red[viciar] o dado, utilizando mais um argumento da função `sample()`: .red[`prob = `]<br/>
<BR/>
* O argumento `prob =` deve ser um .red[vetor de comprimento igual ao número de categorias] no objeto que alimenta o argumento `x = ` (no nosso caso, o vetor `dado`); este vetor deve conter .red[probabilidades entre 0 e 1] para cada categoria, de forma que o .red[somatório final seja 1]!

```{r}
dado <- 1:6
n_de_arremessos <- 10
probabilidades <- c(0.29, 0.29, 0.22, 0.11, 0.05, 0.04)
sum(probabilidades)
table(sample(x = dado, size = n_de_arremessos, replace = TRUE, prob = probabilidades))
```

---

# .blue[**Arremessos de dado viciado**]

<BR/>

- Agora podemos reescrever a função para tornar possível viciar os arremessos do dado 

- Podemos incluir a função .red[`table()`] para que a função retorne tanto os .red[valores brutos], quanto a sua .red[frequência de ocorrência]

- Para isso, precisamos armazenar os resutlado em dois objetos diferentes e transformá-los em uma .red[lista]

```{r}
# nossa função
arremecar_dado <- function(numero_de_arremessos) {
  dado <- 1:6
  resultado <- sample(x = dado, size = numero_de_arremessos,
                      replace = TRUE)
  return(resultado)
}
```

---

# .blue[**Arremessos de dado viciado**]


```{r arremesso de dado viciado 2}
# arremesso de dado viciado
 arremecar_dado <- function(numero_de_arremessos, probabilidades) {
  dado <- 1:6
  resultado1 <- sample(x = dado, size = numero_de_arremessos,
                      replace = TRUE, probabilidades)
  resultado2 <- table(sample(x = dado, size = numero_de_arremessos,
                      replace = TRUE, probabilidades))
  return(list(Bruto = resultado1, Frequencias = resultado2))
}

# testar a função
set.seed(100)
arremecar_dado(numero_de_arremessos = 33, probabilidades = probabilidades)
```



???


# .blue[**Arremessos de dado - argumentos default**]

- É interessante que o nosso dado seja .red[honesto] .red[a priore] 

- Para isso podemos determinar um valor seja .red[*default*] para o argumento .red[`probabilidades`]

```{r arremesso de dado viciado 3}
# arremesso de dado viciado
 arremecar_dado <- function(numero_de_arremessos, probabilidades = c(1/6, 1/6, 1/6, 1/6, 1/6, 1/6)) {
  dado <- 1:6
  resultado1 <- sample(x = dado, size = numero_de_arremessos,
                      replace = TRUE, probabilidades)
  resultado2 <- table(sample(x = dado, size = numero_de_arremessos,
                      replace = TRUE, probabilidades))
  return(list(Bruto = resultado1, Frequencias = resultado2))
}

# testar a função
set.seed(100)
arremecar_dado(numero_de_arremessos = 33)
```

???


---

# .blue[**Analisando diferentes planilhas**]

<BR/>

```{r exemplo 1, eval = FALSE}
df1_bruto <- read.table("df1.txt", header = TRUE)
```

```{r, eval = FALSE}
df_limpo1 <- df_bruto1[, c('n_sp', 'area', 'isolamento')]
```

???
Imagine que você precisa analisar uma planilha com dados de número de de espécies, área e isolamento.

---

# .blue[**Analisando diferentes planilhas**]

```{r, eval = FALSE}
df_bruto1 <- read.table("df1.txt", header = TRUE)
df_limpo1 <- df_bruto1[, c('n_sp', 'area', 'isolamento')]
```
--
```{r, eval = FALSE}
df_bruto2 <- read.table("df2.txt", header = TRUE)
df_limpo2 <- df_bruto2[, c('n_sp', 'area', 'isolamento')]

df_bruto3 <- read.table("df3.txt", header = TRUE)
df_limpo3 <- df_bruto3[, c('n_sp', 'area', 'isolamento')]

df_bruto4 <- read.table("df4.txt", header = TRUE)
df_limpo4 <- df_bruto4[, c('n_sp', 'area', 'isolamento')]
```

???

---

# .blue[**Analisando diferentes planilhas**]

```{r, eval = FALSE}
df_bruto1 <- read.table("df1.txt", header = TRUE)
df_limpo1 <- df_bruto1[, c('n_sp', 'area', 'isolamento')]
df_limpo1$isolamento <- as.factor(isolamento)
```
--
```{r, eval = FALSE}
df_bruto2 <- read.table("df2.txt", header = TRUE)
df_limpo2 <- df_bruto2[, c('n_sp', 'area', 'isolamento')]
df_limpo2$isolamento <- as.factor(df_limpo3$isolamento)

df_bruto3 <- read.table("df3.txt", header = TRUE)
df_limpo3 <- df_bruto3[, c('n_sp', 'area', 'isolamento')]
df_limpo3$isolamento <- as.factor(df_limpo3$isolamento)

df_bruto4 <- read.table("df4.txt", header = TRUE)
df_limpo4 <- df_bruto4[, c('n_sp', 'area', 'isolamento')]
df_limpo4$isolamento <- as.factor(df_limpo4$isolamento)
```

???

---

# .blue[**Analisando diferentes planilhas**]

```{r, eval = FALSE}
df_bruto1 <- read.table("df1.txt", header = TRUE)
df_limpo1 <- df_bruto1[, c('n_sp', 'area', 'isolamento')]
df_limpo1$isolamento <- as.factor(isolamento)
df_limpo1 <- df_bruto1[!is.na(n_sp), ]
```
--
```{r, eval = FALSE}
df_bruto2 <- read.table("df2.txt", header = TRUE)
df_limpo2 <- df_bruto2[, c('n_sp', 'area', 'isolamento')]
df_limpo2$isolamento <- as.factor(df_limpo3$isolamento)
df_limpo2 <- df_bruto2[is.na(n_sp), ]

df_bruto3 <- read.table("df3.txt", header = TRUE)
df_limpo3 <- df_bruto3[, c('n_sp', 'area', 'isolamento')]
df_limpo3$isolamento <- as.factor(df_limpo3$isolamento) 
df_limpo3 <- df_bruto3[!is.na(n_sp), ]

df_bruto4 <- read.table("df4.txt", header = TRUE)
df_limpo4 <- df_bruto4[, c('n_sp', 'area', 'isolamento')]
df_limpo4$isolamento <- as.factor(df_limpo4$isolamento)
df_limpo4 <- df_bruto4[!is.na(n_sp), ]
```

???

---

# .blue[**Analisando diferentes planilhas**]

```{r, eval = FALSE}
df_bruto1 <- read.table("df1.txt", header = TRUE)
df_limpo1 <- df_bruto1[, c('n_sp', 'area', 'isolamento')]
df_limpo1$isolamento <- as.factor(isolamento)
df_limpo1 <- df_bruto1[!is.na(n_sp), ]

df_bruto2 <- read.table("df2.txt", header = TRUE)
df_limpo2 <- df_bruto2[, c('n_sp', 'area', 'isolamento')]
df_limpo2$isolamento <- as.factor(df_limpo3$isolamento)
*df_limpo2 <- df_bruto2[is.na(n_sp), ]

df_bruto3 <- read.table("df3.txt", header = TRUE)
df_limpo3 <- df_bruto3[, c('n_sp', 'area', 'isolamento')]
df_limpo3$isolamento <- as.factor(df_limpo3$isolamento) 
df_limpo3 <- df_bruto3[!is.na(n_sp), ]

df_bruto4 <- read.table("df4.txt", header = TRUE)
df_limpo4 <- df_bruto4[, c('n_sp', 'area', 'isolamento')]
df_limpo4$isolamento <- as.factor(df_limpo4$isolamento)
df_limpo4 <- df_bruto4[!is.na(n_sp), ]
```

???

---

# .blue[**Convertendo scripts em funções**]

```{r, eval = FALSE}
df_bruto1 <- read.table("df1.txt", header = TRUE)
df_limpo1 <- df_bruto1[, c('n_sp', 'area', 'isolamento')]
df_limpo1$isolamento <- as.factor(isolamento)
df_limpo1 <- df_bruto1[!is.na(n_sp), ]

df_bruto2 <- read.table("df2.txt", header = TRUE)
df_limpo2 <- df_bruto2[, c('n_sp', 'area', 'isolamento')]
df_limpo2$isolamento <- as.factor(df_limpo3$isolamento)
*df_limpo2 <- df_bruto2[is.na(n_sp), ]

df_bruto3 <- read.table("df3.txt", header = TRUE)
df_limpo3 <- df_bruto3[, c('n_sp', 'area', 'isolamento')]
df_limpo3$isolamento <- as.factor(df_limpo3$isolamento) 
df_limpo3 <- df_bruto3[!is.na(n_sp), ]

df_bruto4 <- read.table("df4.txt", header = TRUE)
df_limpo4 <- df_bruto4[, c('n_sp', 'area', 'isolamento')]
df_limpo4$isolamento <- as.factor(df_limpo4$isolamento)
df_limpo4 <- df_bruto4[!is.na(n_sp), ]
```

???

---

# .blue[**Convertendo scripts em funções**]

<BR/>

1. Prepare um .red["modelo"]<br/>
<BR/>
2. .red[Cole] o script no interior do "modelo"<br/>
<BR/>
3. Escolha os .red[argumentos]<br/>
<BR/>
4. .red[Substitua] os objetos internos pelos .red[argumentos equivalentes]<br/>
<BR/> 
5. Use .red[nomes generalistas] para as variáveis internas a função<br/>
<BR/> 
6. .red[Teste] a função!

???


---


# .blue[**Prepare um "modelo"**]

```{r exemplo 2, eval = FALSE}
importar_dados <- function() {
  
  
  
  

}
```

???


---

# .blue[**Cole o script**]

```{r, eval = FALSE}
importar_dados <- function() {
df_bruto1 <- read.table("df1.txt", header = TRUE)
df_limpo1 <- df_bruto1[, c('n_sp', 'area', 'isolamento')]
df_limpo1$isolamento <- as.factor(isolamento)
df_limpo1 <- df_bruto1[!is.na(n_sp), ]
}
```

???


---

# .blue[**Escolha os argumentos**]

```{r, eval = FALSE}
*importar_dados <- function(nome_do_arquivo) { # somente 1 argumento
df_bruto1 <- read.table("df1.txt", header = TRUE)
df_limpo1 <- df_bruto1[, c('n_sp', 'area', 'isolamento')]
df_limpo1$isolamento <- as.factor(isolamento)
df_limpo1 <- df_bruto1[!is.na(n_sp), ]
}
```

???


---

# .blue[**Substitua pelos argumentos equivalentes**]

```{r, eval = FALSE}
importar_dados <- function(nome_do_arquivo) { 
*df_bruto1 <- read.table(nome_do_arquivo, header = TRUE) # substituir
# pelo argumento equivalente
df_limpo1 <- df_bruto1[, c('n_sp', 'area', 'isolamento')]
df_limpo1$isolamento <- as.factor(isolamento)
df_limpo1 <- df_bruto1[!is.na(n_sp), ]
}
```

???


---

# .blue[**Use nomes generalistas para as variáveis**]

```{r, eval = FALSE}
importar_dados <- function(nome_do_arquivo) { 
*df_bruto <- read.table(nome_do_arquivo, header = TRUE) # nome
# generalista
*df_limpo <- df_bruto[, c('n_sp', 'area', 'isolamento')] # nome
# generalista
df_limpo$isolamento <- as.factor(isolamento)
df_limpo <- df_bruto[!is.na(n_sp), ]
}
```

# .blue[**Teste da função**]

```{r, eval = FALSE}
df1 <- importar_dados("df1.txt")
df2 <- importar_dados("df2.txt")
df3 <- importar_dados("df3.txt")
df4 <- importar_dados("df4.txt")
```

???


---
background-image: url("/home/cla/Documentos/Vitor/Laboratórios/UNIRIO/Disciplinas/Analise-de-Dados-Ecologicos/libs/images/AnalisaR_teste1.png")
background-position: 0% 100%
background-size: 20%
class: center, middle

# .blue[**Agora é sua vez de criar uma função para arremessar moedas**]

---
layout: true
name: basic
background-image: url("/home/cla/Documentos/Vitor/Laboratórios/UNIRIO/Disciplinas/Analise-de-Dados-Ecologicos/libs/images/AnalisaR_teste1.png")
background-position: 0% 100%
background-size: 20%
class: inverse

```{r setup 2, include=FALSE}
options(htmltools.dir.version = FALSE)
library(tidyverse)
df <- read.table("/home/cla/Documentos/Vitor/Laboratórios/UNIRIO/Disciplinas/Introdução-ao-R-para-Ecólogos/IRE-21/Dados/df.txt", header = TRUE)
xaringanExtra::use_xaringan_extra()
xaringanExtra::use_clipboard()
```
---

# .blue[**Controladores de Fluxo**]

<BR/>

* São .red[funções] usadas especialmente para .red[controlar o fluxo] de computação de certas tarefas. <br/>
<BR/>
* Fornecem mais .red[flexibilidade] ao programarmos uma tarefa, especialmente tarefas .red[repetitivas]. <br/>
<BR/>
* Os principais .red[controladores de fluxo] para funções são:<br/>
<BR/>

---

# .blue[**Controladores de Fluxo**]

```{r tabela controladores de fluxo, eval = TRUE, echo = FALSE}
`Controlador de Fluxo` <- c("if, else", "for", "while", "repeat", "break", "next", "return")
Uso <- c("testa uma condição para 
executar uma ação", "executa uma tarefa repetitivamente por um 
número determinado de vezes
", "executa um loop enquanto uma
condição for verdadeira
", "executa uma ação 
infinitamente 
", "interrompe a execução de 
um loop", "pula a iteração de um
loop", "fornecer o output que ela
deve retornar")

knitr::kable(data.frame(`Controlador de Fluxo`  = `Controlador de Fluxo` , Uso = Uso), format = "html", align = "c", caption = "Controladores de fluxo e seus usos no R") %>% 
  kableExtra::column_spec(1:2, background = "black") %>%
  kableExtra::row_spec(0, background = "black")

```

---

## .blue[**`if() {}`, `else() {}`**]

<BR/>

* Esta função executa uma ação se uma .red[condição] for atendida, caso contrário, ela realizará uma outra ação.<br/>
<BR/>  
* Funcionamento básico desta função é:<br/>  
<BR/>
```{r if else, eval = FALSE, echo = TRUE}
if(condição) {
        <ação>
        } else {
        <outra ação>        
        }
```
--
count: false
<BR/>
- OBS: Não confundir com a função .red[`ifelse()`]!

???
Explicar a função `ifelse()`

---

## .blue[**`if() {}`, `else() {}`**]

<BR/>

* Podemos utilizar o exemplo abaixo para entender a funcionalidade desta função:<br/>
<BR/>
```{r exemplo if else 1, eval = FALSE, echo = TRUE}
# vamos criar um objeto com o número 5
x <- 5
# e vamos usar a estrutura de controle if, else
teste_condicional <- if(x < 10){
  "número menor que 10"
} else {
  "número é maior que 10"
}

teste_condicional
```

---

## .blue[**`if() {}`, `else() {}`**]

<BR/>

* Vamos testar com outro valor de .red[`x`]:<br/>
<BR/>

```{r exemplo if else 2, eval = FALSE, echo = TRUE}
# vamos criar um objeto com o número 20
x <- 20
# e vamos usar a estrutura de controle if, else
teste_condicional <- if(x < 10){
  "número menor que 10"
} else {
  "número é maior que 10"
}
teste_condicional
```

---

## .blue[**`if() {}`, `else() {}`**]

<BR/>

* .red[IMPORTANTE]: este controlador de fluxo não está restrito à uma única condição, podendo ser incluídas .red[múltiplas condições].<br/>
<BR/>

```{r exemplo if else 3, eval = TRUE, echo = TRUE}
# vamos criar um objeto com o número 10
x <- 10
# e vamos usar a estrutura de controle if, else
teste_condicional <- if(x < 10){
  "número menor que 10"
} else if(x == 10) {
  "número é igual a 10"
} else {
  "número é maior que 10"
}
teste_condicional
```

---

# .blue[**Aplicação - Argumento default lógico**]

<BR/>

- Podemos utilizar o controlador de fluxo condicional para .red[determinar] como o .red[resultado] da função .red[`arremessar_dado()`] será .red[apresentado]: na .red[forma bruta], na forma de .red[frequência], ou os .red[dois resultados] ao mesmo tempo

---

# .blue[**Aplicação - Argumento default lógico**]

```{r, eval = FALSE}
# arremesso de dado honesto com controle parcial de resultados
arremecar_dado <- function(numero_de_arremessos, 
                           probabilidades = c(1/6, 1/6, 1/6, 1/6, 1/6, 1/6), 
                           resu = TRUE) {
  dado <- 1:6
  resultado1 <- sample(x = dado, size = numero_de_arremessos,
                       replace = TRUE, probabilidades)
  resultado2 <- table(sample(x = dado, size = numero_de_arremessos,
                             replace = TRUE, probabilidades))
  resultados <- list(Bruto = resultado1, Frequencias = resultado2)
  if(resu == TRUE) {
    return(resultados$Frequencias)
  } else {
    return(resultados$Bruto)
  }
}

# testar a função
set.seed(100)
arremecar_dado(numero_de_arremessos = 33)

```

---

# .blue[**Aplicação - Argumento default categórico**]

```{r, eval = FALSE}
# arremesso de dado honesto com controle total de resultados
arremecar_dado <- function(numero_de_arremessos, 
                           probabilidades = c(1/6, 1/6, 1/6, 1/6, 1/6, 1/6), 
                           resu = c("Bruto", "Frequencias", "Ambos")) {
  dado <- 1:6
  resultado1 <- sample(x = dado, size = numero_de_arremessos,
                       replace = TRUE, probabilidades)
  resultado2 <- table(sample(x = dado, size = numero_de_arremessos,
                             replace = TRUE, probabilidades))
  resultados <- list(Bruto = resultado1, Frequencias = resultado2)
  resu <- match.arg(resu)
  if(resu == "Frequencias") {
    return(resultados$Frequencias)
  } else if(resu == "Bruto") {
    return(resultados$Bruto)
  } else {
    return(resultados)
  }
}

# testar a função
set.seed(100)
arremecar_dado(numero_de_arremessos = 33)

```
---

## .blue[**`if() {}`,  `else {}`**]

<BR/>

1. Transforme o teste condicional do exemplo anterior em uma função.<br/>
<BR/>
2. Usando o controlador de fluxo `if`, `else` construa um teste lógico que permita optar pela apresentação do resultado da função arremecar_moeda em uma tabela, sem estar tabelado, ou os dois resultados ao mesmo tempo.<br/>

---

## .blue[**`for() {}`**]

<BR/>
* O .red[`for` executa uma tarefa repetidamente] por um número determinado de vezes<br/>
<BR/>
* Este é o controlador de fluxo mais útil e um dos mais usados ao programarmos funções<br/>
<BR/>
* Sua estrutura básica é a seguinte:<br/>
<BR/>
```{r for() {}, eval = FALSE, echo = TRUE}
for(i in quantidade_de_vezes) { # onde i 
        # é cada uma das vezes
        <ação>
}
```

---

# .blue[**`for() {}`**]

<BR/>

* Um exemplo do uso desse controlador de fluxo é:<br/>
<BR/>
```{r exemplo for 1, eval = FALSE, echo = TRUE}
for (i in 1:10) { # para cada i de 1 a 10
  print(i) # imprima o valor de i
  Sys.sleep(time = 1) # aguarde 1 segundo antes de
  # prosseguir
  # repita a operaçao
}
```


???
* O poder do `for() {}` está na sua capacidade em realizar uma tarefa repetitiva, de forma que a cada vez que essa tarefa é realizada, o seu resultado seja **armazenado** em algum objeto do R.<br/>
<BR/>
* Exemplos desta aplicação são:<br/>
<BR/>
  * Juntar conjuntos de dados que estejam separados em diferentes arquivos em um único arquivo;  
  * Realizar testes de aleatorização;  
  * Usar a informação em uma célula de um conjunto de dados para buscar uma determina informação e adicionar a um vetor/conjunto de dados (e.g., verificar se os nomes de uma espécie batem com a base do GBIF, extrair informações de coordenadas de um mapa,...); 
  * ... (toda e qualquer tarefa que envolva o uso de um mesmo conjunto de ações repetitivas para partes dos dados)

---

## .blue[**`for() {}`**]

<BR/>

* Um exemplo prático: criar uma .red[sequência Fibonacci]<br/>
<BR/>
* Os .red[dois primeiros números] da sequência são 1<br/>
<BR/>
* A partir do terceiro número, .red[todo número é a soma dos dois números anteriores]<br/>
<BR/>
* Exemplo: .red[1, 1, 2, 3, 5, 8, 13, 21,...]<br/>
<BR/>

---

## .blue[**`for() {}`**]

* Para tentar reproduzir essa sequência, vamos primeiro .red[criar um vetor] com os .red[dois] primeiros números .red[`1`'s]:
```{r sequencia de Fibonacci 1,  eval = TRUE, echo = TRUE}
vetor1 <- c(1, 1)
```

* Este é um vetor de comprimento 2 (teste: `length(vetor1)`) e, se .red[indexarmos] a posição .red[3] deste vetor, obteremos um NA:
```{r sequencia de Fibonacci 2, eval = TRUE, echo = TRUE}
# o que acontece se indexarmos a segunda posição deste
# vetor?
vetor1[3]
```

* O comprimento de um vetor não é algo fixo após ele ser criado - novos elementos podem ser adicionados de acordo com nossa necessidade
* Podemos usar o .red[`for() {}`] para que a cada posição adicionada em um vetor seja preenchida pela soma dos valores das duas últimas posições

---

## .blue[**`for() {}`**]

<BR/>

- .red[Some] a .red[primeira] e a .red[segunda] posição do `vetor1` através da indexação.<br/>
<BR/>
```{r}
vetor1[1] + vetor1[2]
```

---

## .blue[**`for() {}`**]

<BR/>

- Agora que já nos lembramos como usar a indexação para fazer a soma, vamos usar o .red[`for() {}`] para criar a sequência Fibonacci. Neste caso, vamos especificar um .red[tamanho] para a sequência (ex. 10), e usar o .red[`for() {}`] para calcular os valores das posições da sequência da .red[posição 3] até aquela que nós determinamos (10). A cada repetição, vamos dizer para o __loop__ armazenar o resultado da soma na nova posição que será adicionada ao .red[`vetor1`].<br/>
<BR/>
```{r}
for(i in 3:10) {
  vetor1[i] <- vetor1[i-1] + vetor1[i-2]
}
vetor1
```

---

## .blue[**`for() {}`**]

<BR/>

- Transforme o seu código em uma função em que possamos .red[escolher o tamanho] da sequência de fibonacci que será gerada.<br/>
<BR/>


---

## .blue[**`for() {}`**]

1. Escreva uma função que calcule a .red[média] de um vetor. Formula da media:
$$\bar{x} = \frac{1}{n} \sum_{i=1}^{n}x_{i}$$<br/>

2. Escreva uma função que calcule a .red[variância] de um vetor. 
Fórmula da variância: 
$$\sigma^{2} = \frac{\sum_{i=1}^{n} 
  \left(x_{i} - \bar{x}\right)^{2}}
  {n-1}$$

3. Escreva uma função que calcule o .red[desvio padrão] de um vetor. Fórmula do desvio padrão: 
$$\sigma = \sqrt{\frac{\sum\limits_{i=1}^{n} \left(x_{i} - \bar{x}\right)^{2}} {n-1}}$$


---

## .blue[**`while() {}`**]

<BR/>

* Executa uma .red[ação] (ou não) enquanto uma .red[condição for verdadeira]<br/>
<BR/>
* A estrutura geral do uso deste controlador de fluxo é:<br/>
<BR/>

```{r while() {}, eval = FALSE, echo = TRUE}
while(condição) {
  <ação>
}
```

---

## .blue[**`while() {}`**]

<BR/>

* Apesar desta estrutura se parecer com a do controlador de fluxo `if`, `else`, ele permite que as .red[condições] sejam .red[testadas] de forma .red[iterativa]<br/>
<BR/>
```{r exemplo while() {} 1, eval = FALSE, echo = TRUE}
contagem  <- 0
while (contagem < 10) {
  print(contagem)
  Sys.sleep(0.5)
  contagem <- contagem + 1
}
```

---

## .blue[**`while() {}`**]

<BR/>

* Este controlador de fluxo pode ser útil se estivermos trabalhando com .red[processo de otimização] em modelagem, e quisermos que o processamento só pare ao atingir uma determinada condição que caracterize uma boa otimização<br/>
<BR/>

* Você também pode combinar este controlador de fluxo com os outros.

---

## .blue[**`while() {}`**]


```{r exemplo while() {} 2, eval = FALSE, echo = TRUE}
# quanto dinheiro tinha na minha conta?
minha_conta <- 0
# como ela vai flutuar?
while(minha_conta < 100 & minha_conta > -100){ 
  # enquanto o que eu tiver na conta estiver em 100 e -100
  print(minha_conta) # imprime o saldo
  
  destino <- sample(x = c("triste", "feliz"), size = 1) 
  # ai entra o destino, como sempre
  if(destino == "feliz") { # se o destino for feliz
  # deposita 20 na minha conta  
    minha_conta <- minha_conta + 20 
  
    } else { # se o destino for triste
    # tira 20 da minha conta
    minha_conta <- minha_conta - 20 
    }
  Sys.sleep(0.5)
}
```

---

## .blue[**`while() {}`**]

* .red[__Dica__]: você pode .red[interromper] o processamento de uma função à qualquer momento, apertando .red[`Esc`] no seu teclado quando o .red[cursor] está no .red[console] ou o botão .red[`Stop`] no console.<br/>
<BR/>

---

# .blue[**`repeat()` + `break()`**]

<BR/>

* O controlador de fluxo .red[`repeat()`] vai .red[repetir] um operação infinitamente, até que você .red[interrompa] o processamento de forma artificial ou até que uma .red[condição] seja atingida e haja a presença do controlador .red[`break()`] - que vai interromper o processamento da função <br/>
<BR/>

```{r exemplo repeat() break(), eval=FALSE, echo = TRUE}
x <- 1
repeat { # aqui está um controlador de fluxo
        print(x)
        Sys.sleep(time = 1)
        x = x+1
        if (x == 6){
                break # aqui está o outro controlador de fluxo
        }

}
```

---

## .blue[**`next()`**]

<BR/>

* Este controlador de fluxo faz com que uma ação seja .red["pulada"], e é útil quando em combinação com outros controladores de fluxo<br/>
<BR/>  

```{r exemplo next(), eval = FALSE, echo = TRUE}
x <- 1:10
for (val in x) {
    if (val > 3 & val < 6){
        next # aqui está o controlador de fluxo
    }
    print(val)
    Sys.sleep(time = 1)
}
```
