==> Slides/07-Visualizacao-de-Dados-com-ggplot2.Rmd <==
---
title: "Introdução ao R para Ecólogos"
subtitle: "⚔<br/>Visualização de Dados com ggplot2"
author: "Vitor Borges"
institute: "PPGBIO-UNIRIO"
date: "2021/06/24"
output:
  xaringan::moon_reader:
    seal: false
    lib_dir: libs
    nature:
      ratio: '16:10'
      slideNumberFormat: ""
      highlightStyle: github
      highlightLines: true
      countIncrementalSlides: false
---

layout: true
name: basic
background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true")
background-position: 0% 100%
background-size: 15%
class: inverse

```{r setup, include=FALSE}
library(dplyr)
library(ggplot2)
library(kableExtra)
library(readr)
source("IRE_functions.R")
theme_set(theme_grey())
options(htmltools.dir.version = FALSE)
xaringanExtra::use_xaringan_extra()
xaringanExtra::use_clipboard()

streams_raw <- readxl::read_xlsx("/home/cla/Documentos/Vitor/Laboratórios/UNIRIO/Disciplinas/Introdução-ao-R-para-Ecólogos/IRE-21/Dados/streams.xlsx")
streams_variables <- read_csv("/home/cla/Documentos/Vitor/Laboratórios/UNIRIO/Disciplinas/Introdução-ao-R-para-Ecólogos/IRE-21/Dados/streams_variables.csv",
                               col_types = cols(
                                 stream = col_factor()
                               ))
streams_date_variables <- read_csv("/home/cla/Documentos/Vitor/Laboratórios/UNIRIO/Disciplinas/Introdução-ao-R-para-Ecólogos/IRE-21/Dados/streams_date_variables.csv",
                               col_types = cols(
                                 stream = col_factor()
                               ))
tadpoles_clean <- read_tadpoles_raw("/home/cla/Documentos/Vitor/Laboratórios/UNIRIO/Disciplinas/Introdução-ao-R-para-Ecólogos/IRE-21/Dados/tadpoles_raw.csv")
tadpoles_clean <- tadpoles_clean %>%
  mutate(ratio = body_len/total_len) 
new_body_len <- tadpoles_clean$total_len[which(tadpoles_clean$ratio > 1)]

new_total_len <- tadpoles_clean$body_len[which(tadpoles_clean$ratio > 1)]

new_body_len[2] <- NA

new_total_len[2] <- NA

tadpoles_clean$total_len[which(tadpoles_clean$ratio > 1)] <- new_total_len

tadpoles_clean$body_len[which(tadpoles_clean$ratio > 1)]  <- new_body_len

tadpoles_clean <- tadpoles_clean %>%
  mutate(ratio = body_len/total_len) 
tadpoles_freq_stream <- tadpoles_clean %>% 
  group_by(stream, species) %>% 
  summarise(n = n()) %>% 
  mutate(freq_rel = round(n/sum(n), 3)) %>% 
  drop_na() 
tadpoles_freq_date_stream <- tadpoles_clean %>% 
  group_by(date = zoo::as.yearmon(date), stream, species) %>% 
  summarise(n = n()) %>% 
  mutate(freq_rel = round(n/sum(n), 3)) %>% 
  drop_na() 
tadpoles_rich_stream <- tadpoles_freq_stream %>% 
  count(stream, name = "n_sp")
tadpoles_rich_date_stream <- tadpoles_freq_date_stream %>%
  group_by(date) %>% 
  count(stream, name = "n_sp")
tadpoles_len_sp <- tadpoles_clean %>% 
  select(date, species, stage, body_len, total_len) %>% 
  mutate(date = zoo::as.yearmon(date),
         ratio = body_len/total_len)
```

---
name: Title
background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true")
background-position: 0% 100%
background-size: 15%
class: inverse, center, middle, title-slide

# .blue[**Introdução ao R para Ecólogos**]

## .blue[**⚔<br/>Visualização de Dados com ggplot2**]

### .blue[Vitor Borges-Júnior]

### .blue[PPGBIO - UNIRIO]

### .blue[2021/07/13]

???
IDEIAS DE MELHORIAS PARA O PRÓXIMO CURSO

- Condensar o conteúdo de visualização dos dados para caber todo em uma aula.

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**ggplot2 - Gramática Gráfica em R**]

.center[.red[**Três Elementos Essenciais da Gramática Gráfica**]]

.pull-left[
```{r, echo = FALSE, fig.align='center', out.width="85%"}
ids <- factor(rep(c("Dados", "Estéticos", "Geométricos","Temas"), 
                  each = 4))

x <- rep(c(2, 1, 3 , 4), 4)

y <- c(0, 0.75, 0.75, 0)

for (i in 5:length(ids)) {
  y[i] <- y[i - 4] + 0.75
}

y1 <- y[2]

for (i in 2:4) {
  y1[i] <- y1[i - 1] + 0.75
}

positions <- data.frame(camada = ids,
                        x = x,
                        y = y)

positions %>% 
  ggplot(aes(x = x,
             y = y, 
             fill = camada)) +
  geom_polygon(alpha  = 0.5,
               show.legend = FALSE) +
  scale_fill_manual(values = c(Dados = "#1B9E77",
                               Estéticos = "#D95F02",
                               Geométricos = "#7570B3",
                               Temas = "#A6761D")) +
  annotate(geom = "text", 
           x = c(0.5, 0.3, 0.125, 0.5), 
           y = y1, 
           label = unique(positions$camada),
           family = "serif",
           fontface = "italic",
           size = 12,
           color = c(Dados = "#1B9E77",
                     Estéticos = "#D95F02",
                     Geométricos = "#7570B3",
                     Temas = "#A6761D")) + 
  xlim(-0.6, 4) +
  ylim(0, 5.5) +
  theme(text = element_blank(),
        line = element_blank(),
        panel.background = element_rect(fill = "black"))
```
]

.pull-right[
<br/>
  <BR/>
<br/>
  <BR/>
```{r ggplot2 elementos essenciais 1, echo = FALSE}
data.frame(Descrição = c("Aparência do gráfico", "O elementos visuais utilizados para representar os dados", "A escala em que estamos mapeando nossos dados", "Os dados que estão sendo plotados")) %>% 
  kbl() %>% 
  kable_material_dark("striped", 
                      fixed_thead = list(enabled = T, 
                                   background = "purple"),
                      full_width = F, 
                      html_font = "Serif") 
```

]

???
Existem três elementos gramaticais essenciais: os dados, os (elementos) estéticos e os (elementos) geométricos.

A camada de dados se referem naturalmente aos dados 

A camada de estéticos se referem a escala a qual mapearemos os dados 

A camada gemoétrica se refere a forma que os dados assumirão no gráfico.

Essas são as camadas obrigatórias que precisamos alimentar com informações para gerar o gráfico

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**ggplot2 - Gramática Gráfica em R**]

.center[.red[**Elementos Essenciais e Não Essenciais da Gramática Gráfica**]]

.pull-left[
<br/>
```{r ggplot2 elementos essenciais e não essenciais 1, echo = FALSE, out.width="85%", fig.align='center'}
ids <- factor(rep(c("Dados", "Estéticos", "Geométricos", "Estatísticas",
                    "Facetas", "Coordenadas","Temas"), 
                  each = 4))

x <- rep(c(2, 1, 3 , 4), 7)

y <- c(0, 1, 1, 0)

for (i in 5:length(ids)) {
  y[i] <- y[i - 4] + 0.75
}

y1 <- y[2]

for (i in 2:7) {
  y1[i] <- y1[i - 1] + 0.75
}

positions <- data.frame(camada = ids,
                        x = x,
                        y = y)

positions %>% 
  ggplot(aes(x = x,
             y = y, 
             fill = camada)) +
  geom_polygon(alpha  = 0.5,
               show.legend = FALSE) +
  scale_fill_manual(values = c(Dados = "#1B9E77",
                               Estéticos = "#D95F02",
                               Geométricos = "#7570B3",
                               Estatísticas = "#E7298A",
                               Facetas = "#66A61E",
                               Coordenadas = "#E6AB02",
                               Temas = "#A6761D")) +
  annotate(geom = "text", 
           x = c(0.5, 0.3, 0.125, 0.2, 0.45, 0.1, 0.5), 
           y = y1, 
           label = unique(positions$camada),
           family = "serif",
           fontface = "italic",
           size = 12,
           color = c(Dados = "#1B9E77",
                     Estéticos = "#D95F02",
                     Geométricos = "#7570B3",
                     Estatísticas = "#E7298A",
                     Facetas = "#66A61E",
                     Coordenadas = "#E6AB02",
                     Temas = "#A6761D")) + 
  xlim(-0.6, 4) +
  theme(text = element_blank(),
        line = element_blank(),
        panel.background = element_rect(fill = "black"))
```
]

.pull-right[
```{r, echo = FALSE, fig.align='center', out.width="85%"}
data.frame(Descrição = c("Aparência do gráfico", "O espaço em que os dados são plotados", "Gera múltiplos gráficos", "Representações que melhoram a compreensão", "Os elementos visuais utilizados para representar os dados", "A escala em que estamos mapeando nossos dados", "Os dados que estão sendo plotados")) %>% 
  kbl() %>% 
  kable_material_dark("striped", 
                      fixed_thead = list(enabled = T, 
                                   background = "purple"),
                      full_width = F, 
                      html_font = "Serif") 
```
]

???
Essas são as camadas não obrigatórias.


---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**Camadas de Estatística**]
<br/>
<BR/>

- .red[Duas] categorias de funções estatísticas

  - Chamadas dentro de um .red[`geom_()`]
  
  - Chamada .red[independentemente]

--

- .red[`stats_()`]


???
Há duas formas de usar a camada estatística. A partir de uma camada geométrica, ou a partir de uma chamada independente.

De uma forma ou de outra, a chamada se dará a partir de uma função `stats_()`.

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**Ligação entre geom\\\_() <-> stat\_()**]

.pull-left[
```{r, eval = FALSE}
p <- tadpoles_len_sp %>% 
  filter(species %in% c("Hp", "Pa", "Sf")) %>%
  drop_na() %>% 
  ggplot(aes(x = body_len))

p +  geom_histogram()
```
]

.pull-right[
```{r, echo = FALSE, warning=FALSE, message=FALSE, fig.align='center'}
p <- tadpoles_len_sp %>% 
  filter(species %in% c("Hp", "Pa", "Sf")) %>%
  drop_na() %>% 
  ggplot(aes(x = body_len))

p +  geom_histogram() +
  theme(text = element_text(size = 24))
```
]

???
Já utilizamos uma função `stat_()` antes quando construímos histogramas. A função `stat_bin()` está sendo chamada juntamente a camada geométrica `geom_histogram()` e fazendo a contagem dentro de cada "caixa".

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**Ligação entre geom\\\_() <-> stat\_()**]

```{r, eval = FALSE}
p + geom_histogram()
p + stat_bin()
```

.pull-left[
```{r, echo = FALSE, warning=FALSE, message=FALSE, fig.align='center', out.width="85%"}
p <- tadpoles_len_sp %>% 
  filter(species %in% c("Hp", "Pa", "Sf")) %>%
  drop_na() %>% 
  ggplot(aes(x = body_len))

p +  geom_histogram() +
  theme(text = element_text(size = 24))
```
]

.pull-right[
```{r, echo = FALSE, warning=FALSE, message=FALSE, fig.align='center', out.width="85%"}
p <- tadpoles_len_sp %>% 
  filter(species %in% c("Hp", "Pa", "Sf")) %>%
  drop_na() %>% 
  ggplot(aes(x = body_len))

p +  stat_bin() +
  theme(text = element_text(size = 24))
```
]

???
O mesmo resultado é obtido usando `geom_histogram()` e `stat_bin()`.

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**Ligação entre geom\\\_() <-> stat\_()**]


```{r, eval = FALSE}
p +  geom_bar(stat = "bin")
```


```{r, echo = FALSE, warning=FALSE, message=FALSE, fig.align='center', out.width="45%"}
p <- tadpoles_len_sp %>% 
  filter(species %in% c("Hp", "Pa", "Sf")) %>%
  drop_na() %>% 
  ggplot(aes(x = body_len))

p +  geom_bar(stat = "bin") +
  theme(text = element_text(size = 24))
```


???
Nas funções `geom_()`, a ligação entre as camadas geométricas e estatísticas se dá pelo argumento `stat = `. 

Note que é possível obter um histograma a partir da função `geom_bar(stat = "bin")`.


---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**Ligação entre geom\\\_() <-> stat\_()**]

```{r, eval = FALSE}
p <- tadpoles_clean %>% 
  filter(species %in% c("Hp", "Pa", "Sf")) %>% 
  select(species, stream) %>% 
  drop_na() %>% 
  ggplot(aes(x = species, fill = stream))

p + geom_bar()
p + stat_count()
```

.pull-left[
```{r, echo = FALSE, warning=FALSE, message=FALSE, fig.align='center', out.width="60%"}
p <- tadpoles_clean %>% 
  filter(species %in% c("Hp", "Pa", "Sf")) %>% 
  select(species, stream) %>% 
  drop_na() %>% 
  ggplot(aes(x = species, fill = stream))

p +  geom_bar() +
  theme(text = element_text(size = 24))
```
]

.pull-right[
```{r, echo = FALSE, warning=FALSE, message=FALSE, fig.align='center', out.width="60%"}
p <- tadpoles_clean %>% 
  filter(species %in% c("Hp", "Pa", "Sf")) %>% 
  select(species, stream) %>% 
  drop_na() %>% 
  ggplot(aes(x = species, fill = stream))

p +  stat_count() +
  theme(text = element_text(size = 24))
```
]

???
No caso da camada geométrica `geom_bar()` a camada estatística equivalente é `stat_count()`.

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**Ligação entre geom\\\_() <-> stat\_()**]

```{r, echo = FALSE, fig.align='center', out.width="85%"}
tibble(`stat_()` = c("stat_bin()", "stat_count()"), `geom_()` = c("geom_histogram()", "geom_bar()")) %>% 
  kbl() %>% 
  kable_material_dark("striped", 
                      fixed_thead = list(enabled = T, 
                                   background = "purple"),
                      full_width = F, 
                      html_font = "Serif") 
```

???
Resumidamente, essa é a ligação entre as camadas geométricas e estatísticas em gráficos de barras.

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**geom\\\_smooth() <-> stat\_smooth()**]

.pull-left[
```{r, eval = FALSE, warning=FALSE, message=FALSE, fig.align='center', out.width="60%"}
p <- tadpoles_clean %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p +  geom_point() +
  geom_smooth() 
```

```
geom_smooth() using method = 'loess' 
and formula 'y ~ x' 
```
]


.pull-right[
```{r, echo = FALSE, warning=FALSE, message=FALSE, fig.align='center'}
p <- tadpoles_clean %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p +  geom_point() +
  geom_smooth() +
  theme(text = element_text(size = 24))
```
]

???
Uma camada geométrica/estatística bastante útil é geom_smooth()/stat_smooth(). 

Elas adicionam linhas de tendência de diferentes tipos.

A região sombreada de cinza ao redor de cada reta representa um intervalo de confiança de 95%.

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**geom\\\_smooth() <-> stat\_smooth()**]

.pull-left[
```{r, eval = FALSE, warning=FALSE, message=FALSE, fig.align='center', out.width="60%"}
p <- tadpoles_clean %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p +  geom_point() +
  geom_smooth(se = FALSE) 
```

```
geom_smooth() using method = 'loess' 
and formula 'y ~ x' 
```
]



.pull-right[
```{r, echo = FALSE, warning=FALSE, message=FALSE, fig.align='center'}
p <- tadpoles_clean %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p +  geom_point() +
  geom_smooth(se = FALSE) +
  theme(text = element_text(size = 24))
```
]

???
É possível desabilitar o intervalo de confiança através do argumento `se = FALSE`.

O tipo da reta que será plotada é controlada pelo argumento `method = `.

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**geom\\\_smooth() <-> stat\_smooth()**]

.pull-left[
```{r, eval = FALSE, warning=FALSE, message=FALSE, fig.align='center', out.width="60%"}
p <- tadpoles_clean %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p +  geom_point() +
  geom_smooth(method = "loess", 
              se = FALSE) 
```

```
geom_smooth() using method = 'loess' 
and formula 'y ~ x' 
```

]

.pull-right[
```{r, echo = FALSE, warning=FALSE, message=FALSE, fig.align='center'}
p <- tadpoles_clean %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p +  geom_point() +
  geom_smooth(se = FALSE) +
  theme(text = element_text(size = 24))
```
]

???
O método *default* é o "loess". 

LOESS é um método não paramétrico de regressão não linear.

O método funciona calculando uma média ponderada de y ao passar uma janela móvel ao longo do eixo x.

É uma ferramenta útil na etapa de exploração dos dados, pois facilita a visualização da forma da relação entre as variáveis.

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**geom\\\_smooth() <-> stat\_smooth()**]

.pull-left[
```{r, eval = FALSE, warning=FALSE, message=FALSE, fig.align='center', out.width="60%"}
p <- tadpoles_clean %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p +  geom_point() +
  geom_smooth(method = "loess", 
              se = FALSE, span = 0.1)
```

```
geom_smooth() using method = 'loess' 
and formula 'y ~ x' 
```
]

.pull-right[
```{r, echo = FALSE, warning=FALSE, message=FALSE, fig.align='center'}
p <- tadpoles_clean %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p +  geom_point() +
  geom_smooth(se = FALSE, span = 0.4) +
  theme(text = element_text(size = 24))
```
]

???
O grau de suavização da curva é controlado pelo argumento `span = ` o qual define o tamanho da janela.

Quanto menor o valor do argumento `span = ` menos linear será a reta.

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**geom\\\_smooth() <-> stat\_smooth()**]

```{r, eval = FALSE, warning=FALSE, message=FALSE, fig.align='center', out.width="60%"}
p <- iris %>%
  ggplot(aes(x = Sepal.Length, 
             y = Sepal.Width,
             color = Species))

p +  geom_point() +
  geom_smooth(method = "loess", 
              se = FALSE, span = 0.4)
```

.pull-left[
.center[.red[`span = 0.4`]]
```{r, echo = FALSE, warning=FALSE, message=FALSE, fig.align='center', out.width="55%"}
p <- iris %>%
  ggplot(aes(x = Sepal.Length, 
             y = Sepal.Width,
             color = Species))

p +  geom_point(show.legend = FALSE) +
  geom_smooth(method = "loess", 
              se = FALSE, span = 0.4) +
  theme(text = element_text(size = 24))
```
]

.pull-right[
.center[.red[`span = 1`]]
```{r, echo = FALSE, warning=FALSE, message=FALSE, fig.align='center', out.width="55%"}
p <- iris %>%
  ggplot(aes(x = Sepal.Length, 
             y = Sepal.Width,
             color = Species))

p +  geom_point(show.legend = FALSE) +
  geom_smooth(method = "loess", 
              se = FALSE, span = 1) +
  theme(text = element_text(size = 24))
```
]

???
Exemplo com outra base de dados para deixar claro o papel do argumento `span = `.

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**geom_smooth(method = "lm")**]

.pull-left[
```{r, eval = FALSE, warning=FALSE, message=FALSE, fig.align='center', out.width="60%"}
p <- tadpoles_clean %>%
  filter(species %in% c("Hp", "Pa", "Sf")) %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p +  geom_point() +
  geom_smooth(method = "lm", 
              se = FALSE)
```
]

.pull-right[
```{r, echo = FALSE, warning=FALSE, message=FALSE, fig.align='center'}
p <- tadpoles_clean %>% 
  filter(species %in% c("Hp", "Pa", "Sf")) %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len,
             color = species))

p +  geom_point() +
  geom_smooth(method = "lm", se = FALSE) +
  theme(text = element_text(size = 24))
```
]

???
O argumento método também pode definir modelos parmétricos, como "lm", como no exemplo, ou "glm", "rlm" e "gam".

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**geom_smooth(method = "lm")**]

.pull-left[
```{r, eval = FALSE, warning=FALSE, message=FALSE, fig.align='center', out.width="60%"}
p <- tadpoles_clean %>%
  filter(species %in% c("Hp", "Pa", "Sf")) %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p +  geom_point() +
  geom_smooth(method = "lm",
              fullrange = TRUE)
```
]

.pull-right[
```{r, echo = FALSE, warning=FALSE, message=FALSE, fig.align='center'}
p <- tadpoles_clean %>% 
  filter(species %in% c("Hp", "Pa", "Sf")) %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len,
             color = species))

p +  geom_point() +
  geom_smooth(method = "lm",
              fullrange = TRUE) +
  theme(text = element_text(size = 24))
```
]

???
Por *default* os modelos são limitados pela variação de cada grupo.

Para métodos paramétricos é possível gerar previsões para além da varaição de cada grupo usando o argumento `fullrange = TRUE`.

Como esperado, o erro padrão aumenta aumenta conforme a reta se afasta dos dados.

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**Ligação entre geom\\\_() <-> stat\_()**]

```{r, echo = FALSE, fig.align='center', out.width="85%"}
tibble(`stat_()` = c("stat_bin()", "stat_count()", "stat_smooth()"), `geom_()` = c("geom_histogram()", "geom_bar()", "geom_smooth()")) %>% 
  kbl() %>% 
  kable_material_dark("striped", 
                      fixed_thead = list(enabled = T, 
                                   background = "purple"),
                      full_width = F, 
                      html_font = "Serif") 
```

???
Em geral, não é necessário utilizar diretamente essas funções estatísticas. 

Elas trabalham nos bastidores da chamada de funções das camadas geométricas. 

Mas é importante conhecer a relação entre as camadas geométricas e suas respectivas camadas estatísticas. 

Isso facilitará a compreensão de mensagens de erro e o uso das páginas de ajuda se você precisar implementar algum ajuste na camada estatística. 

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**Gráficos com Estatísticas**]

.pull-left[
```{r, eval = FALSE, warning=FALSE, message=FALSE, fig.align='center', out.width="60%"}
p <- tadpoles_clean %>%
  filter(species %in% c("Hp", "Pa", "Sf")) %>% 
  drop_na() %>% 
  ggplot(aes(x = species, 
             y = body_len))

p + geom_jitter(width = 0.25, #<<
              seed = 136, 
              alpha = 0.3)
```
]

.pull-right[
```{r, echo = FALSE, warning=FALSE, message=FALSE, fig.align='center'}
p <- tadpoles_clean %>%
  filter(species %in% c("Hp", "Pa", "Sf")) %>% 
  drop_na() %>% 
  ggplot(aes(x = species, 
             y = body_len))

p + geom_jitter(width = 0.25, #<<
              seed = 136, 
              alpha = 0.3) +
  theme(text = element_text(size = 24))
```
]

???
Outra forma de representar esses dados visualmente é calcular e apresentar a média e o desvio padrão, ou o intervalo de confiança de 95%.

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**Gráficos com Estatísticas - geom_pointrange()**]

.center[.red[**Calcular na "mão" **]]
```{r, eval = FALSE}
tadpoles_clean %>%
  filter(species %in% c("Hp", "Pa", "Sf")) %>% 
  drop_na() %>% 
  group_by(species) %>% 
  summarise(media = mean(body_len),
            `+sd` = media + sd(body_len),
            `-sd` = media - sd(body_len))
```

```{r, echo = FALSE}
tadpoles_clean %>%
  filter(species %in% c("Hp", "Pa", "Sf")) %>% 
  drop_na() %>% 
  group_by(species) %>% 
  summarise(media = mean(body_len),
            `+sd` = media + sd(body_len),
            `-sd` = media - sd(body_len)) %>% 
  kbl() %>% 
  kable_material_dark("striped", 
                      fixed_thead = list(enabled = T, 
                                   background = "purple"),
                      full_width = F, 
                      html_font = "Serif") 
  
```


???
Podemos calcular as estatísticas na mão para então plotar no gráfico usando a camada geométrica `geom_pointrange()`.

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**Gráficos com Estatísticas - geom_pointrange()**]

.center[.red[**Calcular na "mão" **]]
.pull-left[
```{r, eval = FALSE}
tadpoles_clean %>%
  filter(species %in% c("Hp", "Pa", "Sf")) %>% 
  drop_na() %>% 
  group_by(species) %>% 
  summarise(media = mean(body_len),
            `+sd` = media + sd(body_len),
            `-sd` = media - sd(body_len)) %>% 
  ggplot(aes(y = media, x = species)) +
  geom_pointrange(aes(ymin = `-sd`,
                      ymax = `+sd`))
```
]

.pull-right[
```{r, echo = FALSE, out.width="95%"}
tadpoles_clean %>%
  filter(species %in% c("Hp", "Pa", "Sf")) %>% 
  drop_na() %>% 
  group_by(species) %>% 
  summarise(media = mean(body_len),
            `+sd` = media + sd(body_len),
            `-sd` = media - sd(body_len)) %>% 
  ggplot(aes(y = media, x = species)) +
  geom_pointrange(aes(ymin = `-sd`, ymax = `+sd`)) +
  theme(text = element_text(size = 24))
  
```
]

???
Podemos calcular as estatísticas na "mão" para então plotar no gráfico usando a camada geométrica `geom_pointrange()`.

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**Gráficos com Estatísticas**]

.center[.red[**`stat_summary()`**]]

.pull-left[
```{r}
library(Hmisc)
smean.sdl(tadpoles_clean$body_len, 
          mult = 1, na.rm = TRUE)
```

```{r, eval = FALSE}
p <- tadpoles_clean %>%
  filter(species %in% c("Hp", "Pa", "Sf")) %>% 
  drop_na() %>% 
  ggplot(aes(x = species, 
             y = body_len))

p + stat_summary(fun.data = mean_sdl,
                 fun.args = list(mult = 1))
```

- Utiliza .red[`geom_pointrange()`] por *default*
]

.pull-right[
```{r, echo = FALSE, out.width="95%"}
p <- tadpoles_clean %>%
  filter(species %in% c("Hp", "Pa", "Sf")) %>% 
  drop_na() %>% 
  ggplot(aes(x = species, 
             y = body_len))

p + stat_summary(fun.data = mean_sdl,
                 fun.args = list(mult = 1)) + 
  theme(text = element_text(size = 24))
  
```
]

???
Ou podemos utilizarcalcular as estatísticas desejadas dentro da função `stat_summary()`. Neste exemplo, usamos a função `mean_sdl()`do pacote `Hmisc` para gerar as estatísticas dentro da função `stat_summary()`. O argumento `mult = 1` é a opção para 1 desvio padrão.

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**Gráficos com Estatísticas**]

.center[.red[**`stat_summary()`**]]

.pull-left[
```{r, eval = FALSE}
p <- tadpoles_clean %>%
  filter(species %in% c("Hp", "Pa", "Sf")) %>% 
  drop_na() %>% 
  ggplot(aes(x = species, 
             y = body_len))

p + stat_summary(fun = mean,
                 geom = "point") +
  stat_summary(fun.data = mean_sdl,
                 fun.args = list(mult = 1),
               geom = "errorbar",
               width = 0.1)
```
]

.pull-right[
```{r, echo = FALSE, out.width="95%"}
p <- tadpoles_clean %>%
  filter(species %in% c("Hp", "Pa", "Sf")) %>% 
  drop_na() %>% 
  ggplot(aes(x = species, 
             y = body_len))

p + stat_summary(fun = mean,
                 geom = "point") +
  stat_summary(fun.data = mean_sdl,
                 fun.args = list(mult = 1),
               geom = "errorbar",
               width = 0.1) + 
  theme(text = element_text(size = 24))
  
```
]

???
Para obter barras de erro basta adicionar uma camada estatística nova, calcular a média e usar o argumento `geom = "point"`.

A segunda camada usar o argumento `geom = "errorbar"` e o argumento `width = ` para determminar a largura da barra.

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**Gráficos com Estatísticas**]

.center[.red[**`stat_summary()`**]]

.pull-left[
```{r, eval = FALSE}
p <- tadpoles_clean %>%
  filter(species %in% c("Hp", "Pa", "Sf")) %>% 
  drop_na() %>% 
  ggplot(aes(x = species, 
             y = body_len))

p + stat_summary(fun = mean,
                 geom = "point") +
  stat_summary(fun.data = mean_cl_normal,
                 fun.args = list(mult = 1),
               geom = "errorbar",
               width = 0.1)
```
]

.pull-right[
```{r, echo = FALSE, out.width="95%"}
p <- tadpoles_clean %>%
  filter(species %in% c("Hp", "Pa", "Sf")) %>% 
  drop_na() %>% 
  ggplot(aes(x = species, 
             y = body_len))

p + stat_summary(fun = mean,
                 geom = "point") +
  stat_summary(fun.data = mean_cl_normal,
               geom = "errorbar",
               width = 0.1) + 
  theme(text = element_text(size = 24))
  
```
]

???
Para obter intervalos de confiança de 95% calculados usando a distribuição-t.

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**Verificando Distribuições**]

```{r, echo = FALSE, message = FALSE, warning=FALSE, fig.align='center', out.width = "50%"}
p <- tadpoles_clean %>%
  filter(species %in% c("Hp", "Pa", "Sf"), 
         total_len < 70,
         body_len < 25) %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p + geom_point() + 
  geom_smooth(method = "lm", se = FALSE, color = "red") +
  theme(text = element_text(size = 24))
  
```

???
O gráfico mostra um modelo de regressão linear ajustado aos dados de tamanho total e corporal de girinos.

Umas das premissas da regressão linear é a normalidade dos resíduos do modelo. 

Isto também implica na normalidade da variável dependente. 

Uma forma de acessar a normalidade de uma variável é a partir de inspeções gráficas.



---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**Verificando Distribuições**]

.center[.red[**`stat_function()`**]]

.pull-left[
```{r, eval = FALSE}
p <- tadpoles_clean %>%
  filter(species %in% c("Hp", "Pa", "Sf"), 
         total_len < 70,
         body_len < 25) %>% 
  drop_na() %>% 
  ggplot(aes(x = body_len))

p + geom_histogram(aes(y = ..density..)) +
  geom_rug() +
  stat_function(fun = dnorm, color = "red",
                args = list(mean = mean(tadpoles_clean$body_len[tadpoles_clean$body_len < 25], na.rm = TRUE),
                            sd = sd(tadpoles_clean$body_len[tadpoles_clean$body_len < 25], na.rm = TRUE)))
```
]

.pull-right[
```{r, echo = FALSE, out.width="95%", warning=FALSE, message=FALSE}
p <- tadpoles_clean %>%
  filter(species %in% c("Hp", "Pa", "Sf"), 
         total_len < 70,
         body_len < 25) %>% 
  drop_na() %>% 
  ggplot(aes(x = body_len))

p + geom_histogram(aes(y = ..density..)) +
  geom_rug() +
  stat_function(fun = dnorm, color = "red",
                args = list(mean = mean(tadpoles_clean$body_len[tadpoles_clean$body_len < 25], na.rm = TRUE),
                            sd = sd(tadpoles_clean$body_len[tadpoles_clean$body_len < 25], na.rm = TRUE))) + 
  theme(text = element_text(size = 24))
  
```
]

???
Usando `stat_function()` é possível especificar qualquer função e pruduzir uma distribuição de probabilidades teórica como uma linha. 

Aqui uma distribuição normal foi produzida utilizando `dnorm()` e seus argumentos, média e desvio padrão, foram adicionados como uma lista para centra-la com base na distribuição empírica dos dados.

Isso permite comparar quão semelhante a dsitrbuição dos dados empíricos é a uma distribuição normal que possui os mesmos parâmetros, ou seja, a mesma média e desvio padrão.

A camada `geom_rug()` serve para mostrar os dados reais que foram encaixotados pelo histograma.

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**Verificando Distribuições**]

.center[.red[**`stat_qq()` + `geom_qq_line()`**]]

.pull-left[
```{r, eval = FALSE}
p <- tadpoles_clean %>%
  filter(species %in% c("Hp", "Pa", "Sf"), 
         total_len < 70,
         body_len < 25) %>% 
  drop_na() %>% 
  ggplot(aes(sample = body_len))

p + stat_qq() +
  geom_qq_line(color = "red")
```
]

.pull-right[
```{r, echo = FALSE, out.width="95%", warning=FALSE, message=FALSE}
p <- tadpoles_clean %>%
  filter(species %in% c("Hp", "Pa", "Sf"), 
         total_len < 70,
         body_len < 25) %>% 
  drop_na() %>% 
  ggplot(aes(sample = body_len))

p + stat_qq() +
  geom_qq_line(color = "red") + 
  theme(text = element_text(size = 24))
  
```
]

???
QQ-plots também são úteis para comparar os dados a uma distribuição.

QQ-plots plotam a amostra contra um distribuição teórica, como uma normal, e desenham uma linha que divide a área de plotagem ao meio. 

Quantos mais os dados se alinham a essa linha, mais próximos eles estão da distribuição teórica.

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**Camada de Facetas**]

<br/>
<BR/>

- Gerar .red[múltiplos] gráficos

- Divisão a partir dos .red[níveis] de uma .red[variável categórica]

- Mesmo .red[sistema de coordendas]

- Facilita a red.[comparação]

???
A camada de facetas gera múltiplos gráficos com base nos diferentes níveis de uma variável categórica.

A ideia é dividir um gráfico que apresenta um grande conjunto de dados em multiplos gráficos que apresentam subconjuntos dos dados originais e que apresentam o mesmo sistema de coordenadas, o que facilita a comparação.

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**Camada de Facetas**]

.center[.red[**Tamanho do Corpo x Tamanho Total x Espécies**]]

.pull-left[
```{r, eval = FALSE}
p <- tadpoles_clean %>%
  filter(species %in% c("Hp", "Pa", "Sf")) %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len,
             color = species))

p + geom_point()
```
]

.pull-right[
```{r, echo = FALSE, out.width="95%", warning=FALSE, message=FALSE}
p <- tadpoles_clean %>%
  filter(species %in% c("Hp", "Pa", "Sf")) %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len,
             color = species))

p + geom_point() +
  theme(text = element_text(size = 24))
```
]

???
Nesse gráfico temos informação sobre a relação entre a tamanho corporal e total de girinos de 3 espécies.

Podemos adicionar uma camada de faceta `facet_grid()` para visualizar a relação individualmente para cada espécie.

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**Camada de Facetas**]

.center[.red[**Tamanho do Corpo x Tamanho Total x Espécies**]]

.pull-left[
```{r, eval = FALSE}
p <- tadpoles_clean %>%
  filter(species %in% c("Hp", "Pa", "Sf")) %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p + geom_point() +
  facet_grid(cols = vars(species))
```
]

.pull-right[
```{r, echo = FALSE, out.width="95%", warning=FALSE, message=FALSE}
p <- tadpoles_clean %>%
  filter(species %in% c("Hp", "Pa", "Sf")) %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p + geom_point() +
  facet_grid(cols = vars(species)) +
  theme(text = element_text(size = 24))
```
]

???
A camada `facet_grid()` tem argumentos para dividir os plots em linhas, ou colunas de acordo com a variável que alimenta a função `var()`.

Aqui, o gráfico foi dividido em colunas usando o argumento `cols = `.

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**Camada de Facetas**]

.center[.red[**Tamanho do Corpo x Tamanho Total x Espécies**]]

.pull-left[
```{r, eval = FALSE}
p <- tadpoles_clean %>%
  filter(species %in% c("Hp", "Pa", "Sf")) %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p + geom_point() +
  facet_grid(. ~ species)
```
]

.pull-right[
```{r, echo = FALSE, out.width="95%", warning=FALSE, message=FALSE}
p <- tadpoles_clean %>%
  filter(species %in% c("Hp", "Pa", "Sf")) %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p + geom_point() +
  facet_grid(. ~ species) +
  theme(text = element_text(size = 24))
```
]

???
O mesmo resultado pode ser obtido a partir da notação de fórmula.

Tudo a direita do til gera gráfcos em diferentes colunas.

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**Camada de Facetas**]

.center[.red[**Tamanho do Corpo x Tamanho Total x Espécies**]]

.pull-left[
```{r, eval = FALSE}
p <- tadpoles_clean %>%
  filter(species %in% c("Hp", "Pa", "Sf")) %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p + geom_point() +
  facet_grid(rows = vars(species))
```
]

.pull-right[
```{r, echo = FALSE, out.width="95%", warning=FALSE, message=FALSE}
p <- tadpoles_clean %>%
  filter(species %in% c("Hp", "Pa", "Sf")) %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p + geom_point() +
  facet_grid(rows = vars(species)) +
  theme(text = element_text(size = 24))
```
]

???
A camada `facet_grid()` tem argumentos para dividir os plots em linhas, ou colunas de acordo com a variável que alimenta a função `var()`.

Aqui, o gráfico foi dividido em colunas usando o argumento `rows = `.

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**Camada de Facetas**]

.center[.red[**Tamanho do Corpo x Tamanho Total x Espécies**]]

.pull-left[
```{r, eval = FALSE}
p <- tadpoles_clean %>%
  filter(species %in% c("Hp", "Pa", "Sf")) %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p + geom_point() +
  facet_grid(species ~ .)
```
]

.pull-right[
```{r, echo = FALSE, out.width="95%", warning=FALSE, message=FALSE}
p <- tadpoles_clean %>%
  filter(species %in% c("Hp", "Pa", "Sf")) %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p + geom_point() +
  facet_grid(species ~ .) +
  theme(text = element_text(size = 24))
```
]

???
Tudo a esquerda do til gera gráfcos em diferentes linhas.

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**Camada de Facetas**]

.center[.red[**Tamanho do Corpo x Tamanho Total x Espécies x Estágio**]]

.pull-left[
```{r, eval = FALSE}
p <- tadpoles_clean %>%
  filter(species %in% c("Hp", "Pa", "Sf")) %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p + geom_point() +
  facet_grid(rows = vars(stage),
             cols= vars(species))
```
]

.pull-right[
```{r, echo = FALSE, out.width="95%", warning=FALSE, message=FALSE}
p <- tadpoles_clean %>%
  filter(species %in% c("Hp", "Pa", "Sf")) %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p + geom_point() +
  facet_grid(rows = vars(stage),
             cols= vars(species)) +
  theme(text = element_text(size = 24))
```
]

???
Subdividindo os gráficos em linhas e colunas a partir da notação moderna.

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**Camada de Facetas**]

.center[.red[**Tamanho do Corpo x Tamanho Total x Espécies**]]

.pull-left[
```{r, eval = FALSE}
p <- tadpoles_clean %>%
  filter(species %in% c("Hp", "Pa", "Sf")) %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p + geom_point() +
  facet_grid(stage ~ species)
```
]

.pull-right[
```{r, echo = FALSE, out.width="95%", warning=FALSE, message=FALSE}
p <- tadpoles_clean %>%
  filter(species %in% c("Hp", "Pa", "Sf")) %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p + geom_point() +
  facet_grid(stage ~ species) +
  theme(text = element_text(size = 24))
```
]

???
Subdividindo os gráficos em linhas e colunas a partir da notação de fórmula.

Problemas típicos ao trabalhar com facetas são os títulos e a ordem dos rótulos. 

Vamos connhecer algumas formas de alterar os títulos e suas ordens.

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**Renomeando Facetas**]

.center[.red[**Tamanho do Corpo x Tamanho Total x Espécies**]]

.pull-left[
```{r, eval = FALSE}
p <- tadpoles_clean %>%
  filter(species %in% c("Hp", "Pa", "Sf")) %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p + geom_point() +
  facet_grid(cols = vars(species), 
             labeller = label_value)
```
]

.pull-right[
```{r, echo = FALSE, out.width="95%", warning=FALSE, message=FALSE}
p <- tadpoles_clean %>%
  filter(species %in% c("Hp", "Pa", "Sf")) %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p + geom_point() +
  facet_grid(cols = vars(species), 
             labeller = label_value) +
  theme(text = element_text(size = 24))
```
]

???
A função *default* da camada faceta que atribui os títulos a partir do argumento `labeller = label_value`.

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**Renomeando Facetas**]

.center[.red[**Tamanho do Corpo x Tamanho Total x Espécies**]]
.pull-left[
```{r, eval = FALSE}
p <- tadpoles_clean %>%
  filter(species %in% c("Hp", "Pa", "Sf")) %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p + geom_point() +
  facet_grid(cols = vars(species), 
             labeller = label_both)
```
]

.pull-right[
```{r, echo = FALSE, out.width="95%", warning=FALSE, message=FALSE}
p <- tadpoles_clean %>%
  filter(species %in% c("Hp", "Pa", "Sf")) %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p + geom_point() +
  facet_grid(cols = vars(species), 
             labeller = label_both) +
  theme(text = element_text(size = 24))
```
]

???
A função mais útil da camada faceta que atribui os títulos a partir do argumento `labeller = label_both` que atribui ao título o nome da variável e o nível.

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**Renomeando Facetas**]

.center[.red[**Renomeando os Dados Diretamente - fct_recode()**]]

.pull-left[
```{r, eval = FALSE}
tadpoles_clean$species <- fct_recode(tadpoles_clean$species,
             `Hylodes pipilans` = "Hp")

p <- tadpoles_clean %>%
  rename(Species = species, Stage = stage) %>% 
  filter(Species %in% c("Hylodes pipilans")) %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p + geom_point() +
  facet_grid(cols = vars(Species, 
                         Stage))
```
]

.pull-right[
```{r, echo = FALSE, out.width="95%", warning=FALSE, message=FALSE}
tadpoles_clean$species <- fct_recode(tadpoles_clean$species,
             `Hylodes pipilans` = "Hp")

p <- tadpoles_clean %>%
  rename(Species = species, Stage = stage) %>% 
  filter(Species %in% c("Hylodes pipilans")) %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p + geom_point() +
  facet_grid(cols = vars(Species, 
                         Stage)) +
  theme(text = element_text(size = 24))
```
]

???
É possível renomear os títulos das facetas diretamente nos dados utilizando a função `fct_recode()` do pacote `forcats`.

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**Renomeando Facetas**]

.center[.red[**Renomeando os Dados Diretamente - fct_recode()**]]


```{r, echo = FALSE, warning=FALSE, message = FALSE}
tadpoles_clean <- read_tadpoles_raw("/home/cla/Documentos/Vitor/Laboratórios/UNIRIO/Disciplinas/Introdução-ao-R-para-Ecólogos/IRE-21/Dados/tadpoles_raw.csv")
```

.pull-left[
```{r, eval = FALSE}
tadpoles_clean$species <- fct_recode(tadpoles_clean$species,
             `Hylodes \npipilans` = "Hp",
             `Proceratophrys \nappendiculata` = "Pa",
             `Scinax \nflavoguttatus` = "Sf")

p <- tadpoles_clean %>%
  rename(Species = species, Stage = stage) %>% 
  filter(Species %in% c("Hylodes \npipilans", 
                        "Proceratophrys \nappendiculata", 
                        "Scinax \nflavoguttatus")) %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p + geom_point() +
  facet_grid(cols = vars(Species))
```
]

.pull-right[
```{r, echo = FALSE, out.width="95%", warning=FALSE, message=FALSE}
tadpoles_clean$species <- fct_recode(tadpoles_clean$species,
             `Hylodes \npipilans` = "Hp",
             `Proceratophrys \nappendiculata` = "Pa",
             `Scinax \nflavoguttatus` = "Sf")

p <- tadpoles_clean %>%
  rename(Species = species, Stage = stage) %>% 
  filter(Species %in% c("Hylodes \npipilans", 
                        "Proceratophrys \nappendiculata", 
                        "Scinax \nflavoguttatus")) %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p + geom_point() +
  facet_grid(cols = vars(Species)) +
  theme(text = element_text(size = 24))
```
]

???
É possível renomear os títulos das facetas diretamente nos dados utilizando a função `fct_recode()` do pacote `forcats`.

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**Reordenando Facetas**]

.center[.red[**Reordenando os Dados Diretamente - fct_relevel()**]]

.pull-left[
```{r, eval = FALSE}
tadpoles_clean$species <- fct_relevel(tadpoles_clean$species,
             "Scinax \nflavoguttatus",
             "Proceratophrys \nappendiculata",
             "Hylodes \npipilans")

p <- tadpoles_clean %>%
  rename(Species = species, Stage = stage) %>% 
  filter(Species %in% c("Hylodes \npipilans", 
                        "Proceratophrys \nappendiculata", 
                        "Scinax \nflavoguttatus")) %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p + geom_point() +
  facet_grid(cols = vars(Species))
```
]

.pull-right[
```{r, echo = FALSE, out.width="95%", warning=FALSE, message=FALSE}
tadpoles_clean$species <- fct_relevel(tadpoles_clean$species,
             "Scinax \nflavoguttatus",
             "Proceratophrys \nappendiculata",
             "Hylodes \npipilans")

p <- tadpoles_clean %>%
  rename(Species = species, Stage = stage) %>% 
  filter(Species %in% c("Hylodes \npipilans", 
                        "Proceratophrys \nappendiculata", 
                        "Scinax \nflavoguttatus")) %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p + geom_point() +
  facet_grid(cols = vars(Species)) +
  theme(text = element_text(size = 24))
```
]

???
É possível alterar a ordem das facetas diretamente nos dados utilizando a função `fct_relevel()` do pacote `forcats`.

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**Camada de Facetas - Ajustando o Espaço de Plotagem**]

.center[.red[**Variáveis Contínuas**]]

.pull-left[
```{r, eval = FALSE}
p <- tadpoles_clean %>%
  rename(Species = species, Stage = stage) %>% 
  filter(Species %in% c("Hylodes \npipilans", 
                        "Proceratophrys \nappendiculata", 
                        "Scinax \nflavoguttatus")) %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p + geom_point() +
  facet_grid(cols = vars(Species),
             rows = vars(Stage))
```
]

.pull-right[
```{r, echo = FALSE, out.width="95%", warning=FALSE, message=FALSE}
p <- tadpoles_clean %>%
  rename(Species = species, Stage = stage) %>% 
  filter(Species %in% c("Hylodes \npipilans", 
                        "Proceratophrys \nappendiculata", 
                        "Scinax \nflavoguttatus")) %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p + geom_point() +
  facet_grid(cols = vars(Species),
             rows = vars(Stage)) +
  theme(text = element_text(size = 24))
```
]

???
Note que os dados de Sf variam menos em relação ao eixo x do que os dados de Hp, gerando muito espaço em branco na área de plotagem dos gráficos de Sf. 

É possível reduzir o espaço em branco permitindo que os eixos x variem em escalas diferentes.

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**Camada de Facetas - Ajustando o Espaço de Plotagem**]

.center[.red[**Variáveis Contínuas - `scales = `**]]

.pull-left[
```{r, eval = FALSE}
p <- tadpoles_clean %>%
  rename(Species = species, Stage = stage) %>% 
  filter(Species %in% c("Hylodes \npipilans", 
                        "Proceratophrys \nappendiculata", 
                        "Scinax \nflavoguttatus")) %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p + geom_point() +
  facet_grid(cols = vars(Species),
             rows = vars(Stage),
             scales = "free_x")
```
]

.pull-right[
```{r, echo = FALSE, out.width="95%", warning=FALSE, message=FALSE}
p <- tadpoles_clean %>%
  rename(Species = species, Stage = stage) %>% 
  filter(Species %in% c("Hylodes \npipilans", 
                        "Proceratophrys \nappendiculata", 
                        "Scinax \nflavoguttatus")) %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p + geom_point() +
  facet_grid(cols = vars(Species),
             rows = vars(Stage),
             scales = "free_x") +
  theme(text = element_text(size = 24))
```
]

???
Isto é alcançado utilizando o argumento `scales = "free_x"`

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**Camada de Facetas - Ajustando o Espaço de Plotagem**]

.center[.red[**Variáveis Contínuas - `scales = `**]]

.pull-left[
```{r, eval = FALSE}
p <- tadpoles_clean %>%
  rename(Species = species, Stage = stage) %>% 
  filter(Species %in% c("Hylodes \npipilans", 
                        "Proceratophrys \nappendiculata", 
                        "Scinax \nflavoguttatus")) %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p + geom_point() +
  facet_grid(cols = vars(Species),
             rows = vars(Stage),
             scales = "free_y")
```
]

.pull-right[
```{r, echo = FALSE, out.width="95%", warning=FALSE, message=FALSE}
tadpoles_clean$species <- fct_recode(tadpoles_clean$species,
             `Hylodes pipilans` = "Hp",
             `Proceratophrys appendiculata` = "Pa",
             `Scinax flavoguttatus` = "Sf")

p <- tadpoles_clean %>%
  rename(Species = species, Stage = stage) %>% 
  filter(Species %in% c("Hylodes \npipilans", 
                        "Proceratophrys \nappendiculata", 
                        "Scinax \nflavoguttatus")) %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p + geom_point() +
  facet_grid(cols = vars(Species),
             rows = vars(Stage),
             scales = "free_y") +
  theme(text = element_text(size = 24))
```
]

???
Também é possível alterar as escalas do eixo y utilizando o argumento `scales = "free_y"`.

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**Camada de Facetas - Ajustando o Espaço de Plotagem**]

.center[.red[**Variáveis Contínuas - `scales = `**]]

.pull-left[
```{r, eval = FALSE}
p <- tadpoles_clean %>%
  rename(Species = species, Stage = stage) %>% 
  filter(Species %in% c("Hylodes \npipilans", 
                        "Proceratophrys \nappendiculata", 
                        "Scinax \nflavoguttatus")) %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p + geom_point() +
  facet_grid(cols = vars(Species),
             rows = vars(Stage),
             scales = "free")
```
]

.pull-right[
```{r, echo = FALSE, out.width="95%", warning=FALSE, message=FALSE}
tadpoles_clean$species <- fct_recode(tadpoles_clean$species,
             `Hylodes pipilans` = "Hp",
             `Proceratophrys appendiculata` = "Pa",
             `Scinax flavoguttatus` = "Sf")

p <- tadpoles_clean %>%
  rename(Species = species, Stage = stage) %>% 
  filter(Species %in% c("Hylodes \npipilans", 
                        "Proceratophrys \nappendiculata", 
                        "Scinax \nflavoguttatus")) %>% 
  drop_na() %>% 
  ggplot(aes(x = total_len, 
             y = body_len))

p + geom_point() +
  facet_grid(cols = vars(Species),
             rows = vars(Stage),
             scales = "free") +
  theme(text = element_text(size = 24))
```
]

???
Também é possível alterar ambas as escalas ao mesmo tempo utilizando o argumento `scales = "free"`.

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**Camada de Facetas - Ajustando o Espaço de Plotagem**]

.center[.red[**Variáveis Categóricas**]]

.pull-left[
```{r, eval = FALSE}
p <- tadpoles_clean %>%
  rename(Species = species, Stage = stage) %>% 
  filter(Species %in% c("Hylodes \npipilans", 
                        "Proceratophrys \nappendiculata", 
                        "Scinax \nflavoguttatus")) %>% 
  drop_na() %>% 
  ggplot(aes(x = body_len, 
             y = Species,
             color = Species))

p + geom_jitter(height = 0.1, 
                show.legend = FALSE) +
  facet_grid(rows = vars(Stage))
```
]

.pull-right[
```{r, echo = FALSE, out.width="95%", warning=FALSE, message=FALSE}
p <- tadpoles_clean %>%
  rename(Species = species, Stage = stage) %>% 
  filter(Species %in% c("Hylodes \npipilans", 
                        "Proceratophrys \nappendiculata", 
                        "Scinax \nflavoguttatus")) %>% 
  drop_na() %>% 
  ggplot(aes(x = body_len, 
             y = Species,
             color = Species))

p + geom_jitter(height = 0.1, 
                show.legend = FALSE) +
  facet_grid(rows = vars(Stage)) +
  theme(text = element_text(size = 24))
```
]

???
Note que Pa não está presente no estágio 42.

---

background-image: url("https://github.com/vntborgesjr/AnalisaR/blob/master/imagens/AnalisaR_teste1.png?raw=true"), url("https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fwww.libraries.rutgers.edu%2Fsites%2Fdefault%2Ffiles%2Fstyles%2Fresize_to_300px_width%2Fpublic%2Fevents%2F2020%2F01%2Fggplot2_0.png%3Fitok%3DMe9kSRI-&f=1&nofb=1")
background-position: 0% 100%, 100% 0%
background-size: 15%, 10%

# .blue[**Camada de Facetas - Ajustando o Espaço de Plotagem**]

.center[.red[**Variáveis Categóricas**]]

.pull-left[
```{r, eval = FALSE}
p <- tadpoles_clean %>%
  rename(Species = species, Stage = stage) %>% 
  filter(Species %in% c("Hylodes \npipilans", 
                        "Proceratophrys \nappendiculata", 
                        "Scinax \nflavoguttatus")) %>% 
  drop_na() %>% 
  ggplot(aes(x = body_len, 
             y = Species,
             color = Species))

p + geom_jitter(height = 0.1, 
                show.legend = FALSE) +
  facet_grid(rows = vars(Stage),
             scales = "free_y",
             space = "free_y")
```
]

.pull-right[
```{r, echo = FALSE, out.width="95%", warning=FALSE, message=FALSE}
p <- tadpoles_clean %>%
  rename(Species = species, Stage = stage) %>% 
  filter(Species %in% c("Hylodes \npipilans", 
                        "Proceratophrys \nappendiculata", 
                        "Scinax \nflavoguttatus")) %>% 
  drop_na() %>% 
  ggplot(aes(x = body_len, 
             y = Species,
             color = Species))

p + geom_jitter(height = 0.1, 
                show.legend = FALSE) +
  facet_grid(rows = vars(Stage),
             scales = "free_y",
             space = "free_y") +
  theme(text = element_text(size = 24))
```
]

???
Podemos controlar o espaço de plotagem do eixo y quando avaliamos variáveis categóricas utilizando o argumento `space = "free_y"`.

- Próximo slide com facet_wrap()

- Camada de coordenadas - apresentar funções
  - coord_fixed()
  - coord_cartesian()
  - xlim()
  - scale_x_continuous(limits = )
  
- Exemplos e diferenças
